feat: implement core user entity, DTOs, and authentication service components

+++ BackEnd/src/main/java/com/lumiedu/user/dto/response/UserResponse.java
@@ -19,8 +19,8 @@ public class UserResponse {
     private UserRole role;
     private AccountStatus accountStatus;
     private Boolean twoFactorEnabled;
-    private Double storageUsedMb;
-    private Double storageLimitMb;
+    private Long storageUsedMb;
+    private Long storageLimitMb;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
